requirements.txt
fastapi==0.98.0
uvicorn==0.22.0
snowflake-connector-python==3.0.0
requests==2.31.0
